# openapi/Infinity.yaml
openapi: 3.0.3
info:
  version: 0.5.3
  title: Infinity API
  description: "A list of the currently provided Interfaces to interact with the infinity, implementing file operations and sending messages"
security:
  - {}
...
servers:
  - url: "http://{apiRoot}:{port}/v1"
    variables:
      apiRoot:
        default: "localhost"
        description: Base address of the local voyager node main API
      port:
        default: 1633
        description: Service port provided in voyager node config
  - url: "http://{apiRoot}:{port}"
    variables:
      apiRoot:
        default: "localhost"
        description: Base address of the local voyager node main API
      port:
        default: 1633
        description: Service port provided in voyager node config
paths:
...
  "/ifi/{reference}":
    get:
      summary: "Get index document from a collection of files"
      tags:
        - Collection
      parameters:
        - in: path
          name: reference
          schema:
            $ref: "InfinityCommon.yaml#/components/schemas/InfinityReference"
          required: true
          description: Infinity address of content
        - $ref: "InfinityCommon.yaml#/components/parameters/InfinityRecoveryTargetsParameter"
      responses:
        "200":
          description: Ok
          headers:
            "infinity-recovery-targets":
              $ref: "InfinityCommon.yaml#/components/headers/InfinityRecoveryTargets"
          # "infinity-feed-index":
          #   $ref: "InfinityCommon.yaml#/components/headers/InfinityFeedIndex"
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "400":
          $ref: "InfinityCommon.yaml#/components/responses/400"
        "404":
          $ref: "InfinityCommon.yaml#/components/responses/404"
        "500":
          $ref: "InfinityCommon.yaml#/components/responses/500"
        default:
          description: Default response
  "/ifi/{reference}/{path}":
    get:
      summary: "Get referenced file from a collection of files"
      tags:
        - Collection
      parameters:
        - in: path
          name: reference
          schema:
            $ref: "InfinityCommon.yaml#/components/schemas/InfinityReference"
          required: true
          description: Infinity address of content
        - in: path
          name: path
          schema:
            type: string
          required: true
          description: Path to the file in the collection.
        - $ref: "InfinityCommon.yaml#/components/parameters/InfinityRecoveryTargetsParameter"
      responses:
        "200":
          description: Ok
          headers:
            "infinity-recovery-targets":
              $ref: "InfinityCommon.yaml#/components/headers/InfinityRecoveryTargets"
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary

        "400":
          $ref: "InfinityCommon.yaml#/components/responses/400"
        "404":
          $ref: "InfinityCommon.yaml#/components/responses/404"
        "500":
          $ref: "InfinityCommon.yaml#/components/responses/500"
        default:
          description: Default response
...
  "/pin/ifi/{address}":
    parameters:
      - in: path
        name: address
        schema:
          $ref: "InfinityCommon.yaml#/components/schemas/InfinityAddress"
        required: true
        description: Infinity address of the collection
    post:
      summary: Pin collection with given address
      tags:
        - Collection pinning
      responses:
        "200":
          description: Pinning collection chunks (and all referenced files) with address
          content:
            application/json:
              schema:
                $ref: "InfinityCommon.yaml#/components/schemas/Response"
        "400":
          $ref: "InfinityCommon.yaml#/components/responses/400"
        "403":
          $ref: "InfinityCommon.yaml#/components/responses/403"
        "404":
          $ref: "InfinityCommon.yaml#/components/responses/404"
        default:
          description: Default response
    delete:
      summary: Unpin file chunks with given address
      tags:
        - Collection pinning
      responses:
        "200":
          description: Unpinning collection chunks (and all referenced files) with address
          content:
            application/json:
              schema:
                $ref: "InfinityCommon.yaml#/components/schemas/Response"
        "400":
          $ref: "InfinityCommon.yaml#/components/responses/400"
        "403":
          $ref: "InfinityCommon.yaml#/components/responses/403"
        "404":
          $ref: "InfinityCommon.yaml#/components/responses/404"
        default:
          description: Default response
...